requirements.txt
numpy==1.20.1
pandas==1.2.2
swat==1.9.3
python-dotenv==0.15.0
scikit-learn==0.24.1
python-slugify==4.0.1
openpyxl==3.0.6
typer==0.3.2
kaggle==1.5.12
pretty-errors==1.2.19
rich==9.9.0
mlflow==1.28.0